Removed basic test class.
Reorganising View Controllers structure, adding a new View Controller that performs data fetching operations, allowing non-TableViewControllers to use the logic for data fetching.
Creating a new structure, to be used when fully tested to automate the decision if the data based view controller should automatically fetch more data based on the last history of the View controller update.
Added the Core data model and entities to be used in the process.

Added the first tests.